pyproject.toml
[project]
name = "tablo"
version = "0.1.0"
description = "A PostGIS table to feature service app for Django"
readme = "readme.rst"
requires-python = "==3.11.*"
dependencies = [
  "Django==4.2.*",  
  "sqlparse>=0.3.1",
  "pyproj<3.7",
  "numpy>=2.4.4, <3.0",
  "cython<3.1",
  "pandas==3.*",
  "django-tastypie==0.15.*",
  "psycopg2-binary",
  "Pillow==7.1.2",
  "django-storages==1.14.*",
  "boto3==1.14.*",
  "sqlalchemy==1.3.*",
  "geoalchemy2==0.7.*",
]

[tool.pytest.ini_options]
pythonpath = ["."]
DJANGO_SETTINGS_MODULE = "tablo.tests.settings"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
]
